Split QuestList into QuestList + QuestListItem
QuestList now only owns the <ul> and the accordion (which single row is
expanded); everything about how a row looks or behaves — collapsed/
expanded rendering, context menu, checklist recurrence scope, reminder
menu, priority cycling, all the due/repeat/timestamp formatting — moves
to a new self-contained QuestListItem, mirroring the pattern
ActiveQuestPanel already uses for the Focus view's single quest.

Updates RecurringChecklistScope.spec.ts's three QuestList mounts to add
QuestListItem: false, since shallowMount auto-stubs it as QuestList's
only direct child otherwise — the QuestEditor/RecurrenceScopeSheet stubs
those tests already configure live two levels deeper now.
